👌 Refactor 0.0.88 to v3
