+++ upgrades/0.0.88.activate-argo-app-sync-v3/context.go
@@ -3,7 +3,7 @@ package main
 import (
 	"context"
 
-	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.88.activate-argo-app-sync-v2/pkg/logger"
+	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.88.activate-argo-app-sync-v3/pkg/logger"
 	"github.com/spf13/afero"
 )
