+++ upgrades/0.0.88.activate-argo-app-sync-v3/main.go
@@ -8,9 +8,9 @@ import (
 	"os"
 	"path/filepath"
 
-	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.88.activate-argo-app-sync-v2/pkg/logger"
+	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.88.activate-argo-app-sync-v3/pkg/logger"
 
-	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.88.activate-argo-app-sync-v2/pkg/commonerrors"
+	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.88.activate-argo-app-sync-v3/pkg/commonerrors"
 	"github.com/oslokommune/okctl/pkg/logging"
 	"github.com/spf13/afero"
 	"github.com/spf13/cobra"
@@ -39,7 +39,7 @@ type cmdFlags struct {
 }
 
 func buildRootCommand() *cobra.Command {
-	err := logging.InitLogger("/tmp/0.0.88.activate-argo-app-sync-v2.log")
+	err := logging.InitLogger("/tmp/0.0.88.activate-argo-app-sync-v3.log")
 	if err != nil {
 		log.Fatal(fmt.Sprintf("error initializing logger: %s", err.Error()))
 	}
