+++ upgrades/0.0.88.activate-argo-app-sync-v3/pkg/argocd/migrate-existing-application-manifests.go
