+++ upgrades/0.0.88.activate-argo-app-sync-v3/pkg/argocd/migrate-existing-application-manifests_test.go
@@ -6,7 +6,7 @@ import (
 	"path"
 	"testing"
 
-	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.88.activate-argo-app-sync-v2/pkg/logger"
+	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.88.activate-argo-app-sync-v3/pkg/logger"
 	"github.com/oslokommune/okctl/pkg/apis/okctl.io/v1alpha1"
 	"github.com/oslokommune/okctl/pkg/jsonpatch"
 	"github.com/oslokommune/okctl/pkg/scaffold"
